Handle /.well-known/ URIs as static files
If the URI is not found in the static files return 404 instead of
delegating to /index.php

Various tools may repeatedly poll or scan for entries under this path,
so don't waste CPU time delegating requests to PHP scripts which will
most likely not have anything to handle them.

This could all change if it emerges that there are plugins which do
handle requests under the /.well-known/ path, in which case it will
probably be made optional.